code/requirements.txt
pydub>=0.25.1
torch>=2.1.0
torchaudio>=2.1.0
whisper-openai>=1.0.0
pygame>=2.5.0
opencv-python>=4.8.0
cairosvg>=2.7.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.65.0
pocketsphinx>=5.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
